MFDv5: fixed overflow problem in XLALFindSmallestValidSamplingRate() - can triggered by high sampling rates and/or large gaps - refs #2162 Original: f68e5639d1145980cddd7ca7c0f8a686164de4be

+++ lalpulsar/src/CWMakeFakeData.c
@@ -504,7 +504,7 @@ XLALFindSmallestValidSamplingRate ( UINT4 *n1,				//< [out] minimal valid sampli
       // now reduce gap to remainder wrt Tsft
       INT4 gap_i = gap_i0 % Tsft;
 
-      if ( (gap_i * nCur) % Tsft == 0 ) {
+      if ( ((INT8)gap_i * nCur) % Tsft == 0 ) {
         continue;
       }
